refactor: restructure frontend directories
