+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import Dashboard from './components/Dashboard';
+import Dashboard from './pages/Dashboard';
 
 function App() {
   return <Dashboard />;
