+++ frontend/src/types/global.d.ts
